Combining UserEdit and SystemWidePermissionsForm
